+++ src/main/resources/application.properties
@@ -143,5 +143,6 @@ spring.banner.location=classpath:ccd-data-store-banner.txt
 ccd.oauth2.client_id: ${OAUTH2_CLIENT_ID:ccd_gateway}
 ccd.oauth2.client_secret: ${OAUTH2_CLIENT_SECRET:OOOOOOOOOOOOOOOO}
 ccd.oauth2.redirect_url: ${OAUTH2_REDIRECT_URL:http://localhost:3451/oauth2redirect}
-ccd.cache.warm.up.email: ${CCD_CACHE_WARM_UP_EMAIL:ccdcachewarmer@gmail.com}
-ccd.cache.warm.up.password: ${CCD_CACHE_WARM_UP_PASSWORD:Monday01!}
+ccd.cache.warm.up.email: ${CCD_CACHE_WARM_UP_EMAIL:andrzej.folga@hmcts.net}
+ccd.cache.warm.up.password: ${CCD_CACHE_WARM_UP_PASSWORD:Monday01}
+ccd.cache.warm.up.sleep.time: ${CCD_CACHE_WARM_UP_SLEEP_TIME_SEC:2}
